requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-telegram-bot
httpx
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
alembic
asyncpg
aiosqlite

